websocket : 1. fix the app.js content. 2. move GreetingWSController.java to the same package of Application. 3. add @Getter & @Setter of GreeingWS class.
