+++ src/main/java/com/bridge/entity/websocket/GreetingWS.java
@@ -1,8 +1,12 @@
 package com.bridge.entity.websocket;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
 
 @AllArgsConstructor
+@Setter
+@Getter
 public class GreetingWS {
     private String content;
 }
